requirements.txt
# Core dependencies
pydantic>=2.0.0
pydantic-settings>=2.0.0
rich>=13.0.0
prompt-toolkit>=3.0.0
typer>=0.9.0
loguru>=0.7.0
httpx>=0.25.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Database
aiosqlite>=0.19.0

# Models (optional - install as needed)
# For Ollama: no additional dependencies
# For Google AI Studio
google-generativeai>=0.3.0

# For Vertex AI
google-cloud-aiplatform>=1.40.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
anyio>=4.0.0